% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_data.R
\name{age_data}
\alias{age_data}
\title{Simulated Age Data}
\format{
A data frame with 400 rows and 1 variable.
}
\source{
Nathanael Gay
}
\description{
Simulated human ages, rounded to the nearest year of life.
\itemize{
\item Age The age of a human rounded to the nearest year of life.
}
}
