% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom}
\alias{binom}
\title{Binomial Probabilities}
\usage{
binom(n, x, p)
}
\arguments{
\item{n}{The number of trials}

\item{x}{Number of successes}

\item{p}{Probability of a success on each trial}
}
\value{
The binomial probability of seeing x successes out of n trials when the probability of a success is p.
}
\description{
A wrapper function for dbinom.
}
