% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12ex6.R
\name{ch12ex6}
\alias{ch12ex6}
\title{14th Edition Triola Data Set}
\format{
A data frame.
}
\source{
Triola, M. 2022. Elementary Statistics, 14th ed. Boston, MA: Pearson.
}
\description{
Consult data set description in question #6 on page 635.
}
