% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12hw1.R
\name{ch12hw1}
\alias{ch12hw1}
\title{Chapter 12.3 Homework#1: Days of Scent Effectiveness}
\format{
A data frame with 12 rows and 3 variables
}
\source{
Nathanael Gay
}
\description{
Days of scent effectiveness by humidity and temperature category.
\itemize{
\item Days
\item Humidity.Level (Category 1 or 2)
\item Temperature (Category 1 or 2)
}
}
