% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12hw10.R
\name{ch12hw10}
\alias{ch12hw10}
\title{14th Edition Triola Data Set}
\format{
A data frame.
}
\source{
Triola, M. 2022. Elementary Statistics, 14th ed. Boston, MA: Pearson.
}
\description{
Consult data set description in question #10 on page 636.
}
