% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12hw2.R
\name{ch12hw2}
\alias{ch12hw2}
\title{Chapter 12.3 Homework#2: Months of Paint Effectiveness}
\format{
A data frame with 40 rows and 3 variables
}
\source{
Nathanael Gay
}
\description{
Months of paint effectiveness by geographic location (United States) and type of paint.
\itemize{
\item Months
\item Location
\item Type
}
}
