% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds45.R
\name{ds45}
\alias{ds45}
\title{14th Edition Triola Data Set}
\format{
A data frame.
}
\source{
Triola, M. 2022. Elementary Statistics, 14th ed. Boston, MA: Pearson.
}
\description{
Consult data set descriptions in the textbook Appendix B, starting on page 749.
}
