% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance.R
\name{insurance}
\alias{insurance}
\title{Insurance: Insured Home Values}
\format{
A data frame with 180,400 rows and 1 variable.
}
\source{
Bethany Kocher
}
\description{
Home insured values from a random sample of Florida properties.
\itemize{
\item value The insured value (in dollars) of the property.
}
}
