% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.plot.R
\name{normal.plot}
\alias{normal.plot}
\title{Plot One or Two Normal Density Functions}
\usage{
normal.plot(mean = 0, sd = 1, mean2 = NA, sd2 = NA)
}
\arguments{
\item{mean}{Mean of the first distribution}

\item{sd}{Standard deviation of the first distribution}

\item{mean2}{Mean of the second distribution}

\item{sd2}{Standard deviation of the second distribution}
}
\value{
Density graphs
}
\description{
Plots the normal density function of one or two graphs.
}
