% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Permutation}
\usage{
permutation(n, r)
}
\arguments{
\item{n}{Number of objects being picked from}

\item{r}{Number of objects to be ordered}
}
\value{
The number of ways to select r objects, in order, from n objects
}
\description{
Permutation
}
