% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.z.R
\name{shade.z}
\alias{shade.z}
\title{Plot and Shade A Normal Denisty Function}
\usage{
shade.z(mean = 0, sd = 1, min = -Inf, max = Inf)
}
\arguments{
\item{mean}{The mean of the data set. The default is 0.}

\item{sd}{The standard deviation of the data set.  The default is 1.}

\item{min}{The x-value of the minimum shaded region.  The default is -Infinity.}

\item{max}{The x-value of the maximum shaded region.  The default is Infinity.}
}
\value{
A shaded normal denisty plot, area of shaded region.
}
\description{
Will plot a normal density function and shade in the selected direction.
}
